moved each component for the respective folder
